+++ BB3/App/gui/map/map_thread.c
@@ -58,5 +58,6 @@ void thread_map_start(void *argument)
 //        taskYIELD();
     }
 
+    INFO("Done");
     osThreadSuspend(thread_map);
 }